Trying to get the pop up to depend on the start date
